+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/message/abstracts/MediaGroupMessage.kt
@@ -3,7 +3,6 @@ package com.github.insanusmokrassar.TelegramBotAPI.types.message.abstracts
 import com.github.insanusmokrassar.TelegramBotAPI.types.MediaGroupIdentifier
 import com.github.insanusmokrassar.TelegramBotAPI.types.message.content.abstracts.MediaGroupContent
 
-interface MediaGroupMessage : Message {
+interface MediaGroupMessage : CommonMessage<MediaGroupContent> {
     val mediaGroupId: MediaGroupIdentifier
-    val content: MediaGroupContent
 }