+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 
 * Change algorithm of `executeUnsafe`: now it use loop instead of recursive calling
 * Add additional `startGettingUpdates` with better management of received updates for media groups
+* Now `MediaGroupMessage` is `CommonMessage` with `MediaGroupContent` content
 
 ## 0.9.0